+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.7.0] - Unreleased
+## [0.7.0] - 2024-11-15
 ### Added
 - Add MQTTNet ManagedClient extension version 4.1.4.563.
 - `StreamData` change to `async StreamData` (BREAKING CHANGE!).